refactor(config): restructure configuration modules and remove retrieval config
- Move indexer, llm_pool, metrics, and storage modules from types/
  directory to src/ directory root
- Remove retrieval module as it's no longer needed in core configuration
- Update lib.rs to export modules directly instead of through types mod
- Add comprehensive Config struct with validation capabilities
- Include ConfigValidationError and ValidationError types for proper
  configuration validation
- Add tests for configuration defaults and validation

feat(blackboard): enhance discovery extraction with cross-document refs

- Extract cross-references from evidence content using regex patterns
- Add "cross_ref" and "lead" discovery types for document references
- Track evidence-referenced documents and generate lead discoveries

refactor(analyzer): consolidate JSON parsing utilities

- Move _parse_json_response function to shared utils module
- Import parse_json_response from vectorless.ask.utils
- Update QueryAnalyzer to use consolidated JSON parsing utility

feat(analyzer): add analysis completion tracking

- Add analysis_complete flag to QueryAnalysis to track whether
  deep analysis stages completed successfully
- Set analysis_complete=False when deep analysis stages fail
- Propagate completion status through re_analyze method

refactor(python): remove deprecated retrieval configuration

- Remove set_top_k and set_max_iterations methods from PyConfig
- Update Config documentation to reflect removal of retrieval params
- Remove RetrievalConfig from python exports

refactor(utils): centralize JSON response parsing logic

- Create parse_json_response utility function in utils module
- Consolidate JSON parsing logic from analyzer and verifier modules
- Handle markdown-wrapped JSON and extract JSON blocks properly

feat(verify): improve evidence reference formatting and scoring

- Update verify prompt to use "doc_name/node_title" format for
  evidence references
- Modify DimensionScore to accept both "doc_name/node_title" and
  "node_title" formats
- Calculate overall confidence from dimension scores instead of
  relying on LLM self-assessment
