+++ vectorless/config/__init__.py
@@ -5,7 +5,6 @@ from vectorless.config.models import (
     EngineConfig,
     LlmConfig,
     MetricsConfig,
-    RetrievalConfig,
     RetryConfig,
     StorageConfig,
     ThrottleConfig,
@@ -15,7 +14,6 @@ __all__ = [
     "EngineConfig",
     "LlmConfig",
     "MetricsConfig",
-    "RetrievalConfig",
     "RetryConfig",
     "StorageConfig",
     "ThrottleConfig",
